TIP-860: Move controller to the right BC
