+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/controllers.yml
@@ -233,3 +233,24 @@ services:
             - { related_entity: product, form_type: 'Akeneo\Pim\Enrichment\Bundle\Form\Type\CategoryType', acl: pim_enrich_product, route: pim_enrich }
         calls:
             - [ setContainer, [ '@service_container' ] ]
+
+    pim_enrich.controller.file:
+        class: 'Akeneo\Pim\Enrichment\Bundle\Controller\Ui\FileController'
+        arguments:
+            - '@liip_imagine.controller'
+            - '@akeneo_file_storage.file_storage.filesystem_provider'
+            - '@akeneo_file_storage.repository.file_info'
+            - '@pim_enrich.guesser.file_type'
+            - '@pim_enrich.provider.default_image'
+            - ['catalogStorage']
+
+    pim_enrich.controller.product_model:
+        class: 'Akeneo\Pim\Enrichment\Bundle\Controller\Ui\ProductModelController'
+        arguments:
+            - '@pim_catalog.repository.product_model'
+            - '@pim_catalog.values_filler.entity_with_family_variant'
+            - '@pim_catalog.repository.category'
+            - '@oro_security.security_facade'
+            - '%pim_catalog.entity.category.class%'
+            - 'pim_enrich_product_model_categories_view'
+            - '@@PimEnrich/ProductModel/listCategories.json.twig'
