+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/routing/ui/media.yml
