+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/routing/ui/product_model.yml
