+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Resources/config/controllers.yml
@@ -86,3 +86,17 @@ services:
             - '@pim_standard_format_serializer'
             - '@akeneo_batch_queue.manager.job_execution_manager'
             - '@pim_enrich.repository.job_execution'
+
+
+    pim_enrich.controller.job_tracker:
+        class: 'Akeneo\Platform\Bundle\ImportExportBundle\Controller\Ui\JobTrackerController'
+        arguments:
+            - '@templating'
+            - '@translator'
+            - '@event_dispatcher'
+            - '@pim_enrich.repository.job_execution'
+            - '@pim_connector.event_listener.archivist'
+            - '@pim_standard_format_serializer'
+            - '@akeneo_batch_queue.manager.job_execution_manager'
+            - '@oro_security.security_facade'
+            - {'import': 'pim_importexport_import_profile_show', 'export': 'pim_importexport_export_profile_show'}
